+++ vm/src/types/mod.rs
@@ -3,5 +3,6 @@ pub mod exec_scope;
 pub mod instance_definitions;
 pub mod instruction;
 pub mod layout;
+pub mod layout_name;
 pub mod program;
 pub mod relocatable;