+++ vm/src/vm/errors/runner_errors.rs
@@ -2,6 +2,7 @@
 #![allow(clippy::explicit_auto_deref)]
 
 use crate::stdlib::{collections::HashSet, prelude::*};
+use crate::types::layout_name::LayoutName;
 use thiserror_no_std::Error;
 
 use super::{memory_errors::MemoryError, trace_errors::TraceError};
@@ -39,9 +40,7 @@ pub enum RunnerError {
     #[error("EcOpBuiltin: point {0:?} is not on the curve")]
     PointNotOnCurve(Box<(Felt252, Felt252)>),
     #[error("Builtin(s) {:?} not present in layout {}", (*.0).0, (*.0).1)]
-    NoBuiltinForInstance(Box<(HashSet<&'static str>, String)>),
-    #[error("Invalid layout {0}")]
-    InvalidLayoutName(Box<str>),
+    NoBuiltinForInstance(Box<(HashSet<&'static str>, LayoutName)>),
     #[error("end_run called twice.")]
     EndRunCalledTwice,
     #[error("end_run must be called before finalize_segments.")]