+++ vm/src/vm/runners/cairo_runner.rs
@@ -7,14 +7,14 @@ use crate::{
         ops::{Add, AddAssign, Mul, MulAssign, Sub, SubAssign},
         prelude::*,
     },
-    types::layout::MEMORY_UNITS_PER_STEP,
+    types::{layout::MEMORY_UNITS_PER_STEP, layout_name::LayoutName},
     vm::{
         runners::builtin_runner::SegmentArenaBuiltinRunner,
         trace::trace_entry::{relocate_trace_register, RelocatedTraceEntry},
     },
+    Felt252,
 };
 
-use crate::Felt252;
 use crate::{
     hint_processor::hint_processor_definition::{HintProcessor, HintReference},
     math_utils::safe_div_usize,
@@ -172,25 +172,20 @@ pub enum RunnerMode {
 impl CairoRunner {
     pub fn new_v2(
         program: &Program,
-        layout: &str,
+        layout: LayoutName,
         mode: RunnerMode,
     ) -> Result<CairoRunner, RunnerError> {
         let cairo_layout = match layout {
-            "plain" => CairoLayout::plain_instance(),
-            "small" => CairoLayout::small_instance(),
-            "dex" => CairoLayout::dex_instance(),
-            "recursive" => CairoLayout::recursive_instance(),
-            "starknet" => CairoLayout::starknet_instance(),
-            "starknet_with_keccak" => CairoLayout::starknet_with_keccak_instance(),
-            "recursive_large_output" => CairoLayout::recursive_large_output_instance(),
-            "all_cairo" => CairoLayout::all_cairo_instance(),
-            "all_solidity" => CairoLayout::all_solidity_instance(),
-            "dynamic" => CairoLayout::dynamic_instance(),
-            name => {
-                return Err(RunnerError::InvalidLayoutName(
-                    name.to_string().into_boxed_str(),
-                ))
-            }
+            LayoutName::plain => CairoLayout::plain_instance(),
+            LayoutName::small => CairoLayout::small_instance(),
+            LayoutName::dex => CairoLayout::dex_instance(),
+            LayoutName::recursive => CairoLayout::recursive_instance(),
+            LayoutName::starknet => CairoLayout::starknet_instance(),
+            LayoutName::starknet_with_keccak => CairoLayout::starknet_with_keccak_instance(),
+            LayoutName::recursive_large_output => CairoLayout::recursive_large_output_instance(),
+            LayoutName::all_cairo => CairoLayout::all_cairo_instance(),
+            LayoutName::all_solidity => CairoLayout::all_solidity_instance(),
+            LayoutName::dynamic => CairoLayout::dynamic_instance(),
         };
         Ok(CairoRunner {
             program: program.clone(),
@@ -219,7 +214,7 @@ impl CairoRunner {
 
     pub fn new(
         program: &Program,
-        layout: &str,
+        layout: LayoutName,
         proof_mode: bool,
     ) -> Result<CairoRunner, RunnerError> {
         if proof_mode {
@@ -358,7 +353,7 @@ impl CairoRunner {
         if !program_builtins.is_empty() && !allow_missing_builtins {
             return Err(RunnerError::NoBuiltinForInstance(Box::new((
                 program_builtins.iter().map(|n| n.name()).collect(),
-                self.layout.name.clone(),
+                self.layout.name,
             ))));
         }
 
@@ -1449,11 +1444,9 @@ impl CairoRunner {
         &self,
         vm: &VirtualMachine,
     ) -> Result<PublicInput, PublicInputError> {
-        let layout_name = self.get_layout().name.as_str();
-
         PublicInput::new(
             &self.relocated_memory,
-            layout_name,
+            self.layout.name.to_str(),
             &vm.get_public_memory_addresses()?,
             self.get_memory_segment_addresses(vm)?,
             self.relocated_trace
@@ -1670,7 +1663,7 @@ mod tests {
     #[cfg_attr(target_arch = "wasm32", wasm_bindgen_test)]
     fn initialize_builtins_with_disordered_builtins() {
         let program = program![BuiltinName::range_check, BuiltinName::output];
-        let cairo_runner = cairo_runner!(program, "plain");
+        let cairo_runner = cairo_runner!(program, LayoutName::plain);
         let mut vm = vm!();
         assert!(cairo_runner.initialize_builtins(&mut vm, false).is_err());
     }
@@ -1679,7 +1672,7 @@ mod tests {
     #[cfg_attr(target_arch = "wasm32", wasm_bindgen_test)]
     fn initialize_builtins_missing_builtins_no_allow_missing() {
         let program = program![BuiltinName::output, BuiltinName::ecdsa];
-        let cairo_runner = cairo_runner!(program, "plain");
+        let cairo_runner = cairo_runner!(program, LayoutName::plain);
         let mut vm = vm!();
         assert_matches!(
             cairo_runner.initialize_builtins(&mut vm, false),
@@ -3828,7 +3821,7 @@ mod tests {
         .unwrap();
 
         let mut hint_processor = BuiltinHintProcessor::new_empty();
-        let mut cairo_runner = cairo_runner!(program, "all_cairo", true);
+        let mut cairo_runner = cairo_runner!(program, LayoutName::all_cairo, true);
         let mut vm = vm!(true);
 
         let end = cairo_runner.initialize(&mut vm, false).unwrap();
@@ -3897,7 +3890,7 @@ mod tests {
             entrypoint: "main",
             trace_enabled: true,
             relocate_mem: false,
-            layout: "all_cairo",
+            layout: LayoutName::all_cairo,
             proof_mode: false,
             secure_run: Some(false),
             ..Default::default()
@@ -3915,7 +3908,7 @@ mod tests {
             entrypoint: "main",
             trace_enabled: false,
             relocate_mem: false,
-            layout: "all_cairo",
+            layout: LayoutName::all_cairo,
             proof_mode: false,
             secure_run: Some(false),
             ..Default::default()
@@ -4034,7 +4027,7 @@ mod tests {
     #[cfg_attr(target_arch = "wasm32", wasm_bindgen_test)]
     fn finalize_segments_run_ended_emptyproof_mode() {
         let program = program!();
-        let mut cairo_runner = cairo_runner!(program, "plain", true);
+        let mut cairo_runner = cairo_runner!(program, LayoutName::plain, true);
         cairo_runner.program_base = Some(Relocatable::from((0, 0)));
         cairo_runner.execution_base = Some(Relocatable::from((1, 0)));
         cairo_runner.run_ended = true;
@@ -4051,7 +4044,7 @@ mod tests {
         Arc::get_mut(&mut program.shared_program_data).unwrap().data =
             vec_data![(1), (2), (3), (4), (5), (6), (7), (8)];
         //Program data len = 8
-        let mut cairo_runner = cairo_runner!(program, "plain", true);
+        let mut cairo_runner = cairo_runner!(program, LayoutName::plain, true);
         cairo_runner.program_base = Some(Relocatable::from((0, 0)));
         cairo_runner.execution_base = Some(Relocatable::from((1, 0)));
         cairo_runner.run_ended = true;
@@ -4085,7 +4078,7 @@ mod tests {
         Arc::get_mut(&mut program.shared_program_data).unwrap().data =
             vec_data![(1), (2), (3), (4)];
         //Program data len = 4
-        let mut cairo_runner = cairo_runner!(program, "plain", true);
+        let mut cairo_runner = cairo_runner!(program, LayoutName::plain, true);
         cairo_runner.program_base = Some(Relocatable::from((0, 0)));
         cairo_runner.execution_base = Some(Relocatable::from((1, 1)));
         cairo_runner.execution_public_memory = Some(vec![1_usize, 3_usize, 5_usize, 4_usize]);
@@ -4190,7 +4183,7 @@ mod tests {
     fn check_range_check_usage_without_builtins() {
         let program = program!();
 
-        let cairo_runner = cairo_runner!(program, "plain");
+        let cairo_runner = cairo_runner!(program, LayoutName::plain);
         let mut vm = vm!();
         vm.builtin_runners = vec![];
         vm.current_step = 10000;
@@ -4512,12 +4505,12 @@ mod tests {
     fn initialize_segments_incorrect_layout_plain_one_builtin() {
         let program = program![BuiltinName::output];
         let mut vm = vm!();
-        let cairo_runner = cairo_runner!(program, "plain");
+        let cairo_runner = cairo_runner!(program, LayoutName::plain);
         assert_eq!(
             cairo_runner.initialize_builtins(&mut vm, false),
             Err(RunnerError::NoBuiltinForInstance(Box::new((
                 HashSet::from([BuiltinName::output.name()]),
-                String::from("plain")
+                LayoutName::plain
             ))))
         );
     }
@@ -4527,12 +4520,12 @@ mod tests {
     fn initialize_segments_incorrect_layout_plain_two_builtins() {
         let program = program![BuiltinName::output, BuiltinName::pedersen];
         let mut vm = vm!();
-        let cairo_runner = cairo_runner!(program, "plain");
+        let cairo_runner = cairo_runner!(program, LayoutName::plain);
         assert_eq!(
             cairo_runner.initialize_builtins(&mut vm, false),
             Err(RunnerError::NoBuiltinForInstance(Box::new((
                 HashSet::from([BuiltinName::output.name(), HASH_BUILTIN_NAME]),
-                String::from("plain")
+                LayoutName::plain
             ))))
         );
     }
@@ -4542,12 +4535,12 @@ mod tests {
     fn initialize_segments_incorrect_layout_small_two_builtins() {
         let program = program![BuiltinName::output, BuiltinName::bitwise];
         let mut vm = vm!();
-        let cairo_runner = cairo_runner!(program, "small");
+        let cairo_runner = cairo_runner!(program, LayoutName::small);
         assert_eq!(
             cairo_runner.initialize_builtins(&mut vm, false),
             Err(RunnerError::NoBuiltinForInstance(Box::new((
                 HashSet::from([BuiltinName::bitwise.name()]),
-                String::from("small")
+                LayoutName::small,
             ))))
         );
     }
@@ -4711,7 +4704,7 @@ mod tests {
         Arc::get_mut(&mut program.shared_program_data).unwrap().data =
             vec_data![(1), (2), (3), (4), (5), (6), (7), (8)];
         //Program data len = 8
-        let mut cairo_runner = cairo_runner!(program, "plain", true);
+        let mut cairo_runner = cairo_runner!(program, LayoutName::plain, true);
         cairo_runner.program_base = Some(Relocatable::from((0, 0)));
         cairo_runner.execution_base = Some(Relocatable::from((1, 0)));
         cairo_runner.run_ended = true;
@@ -4735,7 +4728,7 @@ mod tests {
         Arc::get_mut(&mut program.shared_program_data).unwrap().data =
             vec_data![(1), (2), (3), (4), (5), (6), (7), (8)];
         //Program data len = 8
-        let mut cairo_runner = cairo_runner!(program, "plain", true);
+        let mut cairo_runner = cairo_runner!(program, LayoutName::plain, true);
         cairo_runner.program_base = Some(Relocatable::from((0, 0)));
         cairo_runner.execution_base = Some(Relocatable::from((1, 0)));
         cairo_runner.run_ended = false;
@@ -4753,7 +4746,7 @@ mod tests {
         Arc::get_mut(&mut program.shared_program_data).unwrap().data =
             vec_data![(1), (2), (3), (4), (5), (6), (7), (8)];
         //Program data len = 8
-        let mut cairo_runner = cairo_runner!(program, "plain", true);
+        let mut cairo_runner = cairo_runner!(program, LayoutName::plain, true);
         cairo_runner.program_base = Some(Relocatable::from((0, 0)));
         cairo_runner.execution_base = Some(Relocatable::from((1, 0)));
         cairo_runner.run_ended = true;
@@ -4772,7 +4765,7 @@ mod tests {
         Arc::get_mut(&mut program.shared_program_data).unwrap().data =
             vec_data![(1), (2), (3), (4), (5), (6), (7), (8)];
         //Program data len = 8
-        let mut cairo_runner = cairo_runner!(program, "all_cairo", true);
+        let mut cairo_runner = cairo_runner!(program, LayoutName::all_cairo, true);
         cairo_runner.program_base = Some(Relocatable::from((0, 0)));
         cairo_runner.execution_base = Some(Relocatable::from((1, 0)));
         cairo_runner.run_ended = true;
@@ -4803,7 +4796,7 @@ mod tests {
         Arc::get_mut(&mut program.shared_program_data).unwrap().data =
             vec_data![(1), (2), (3), (4), (5), (6), (7), (8)];
         //Program data len = 8
-        let mut cairo_runner = cairo_runner!(program, "all_cairo", true);
+        let mut cairo_runner = cairo_runner!(program, LayoutName::all_cairo, true);
         cairo_runner.program_base = Some(Relocatable::from((0, 0)));
         cairo_runner.execution_base = Some(Relocatable::from((1, 0)));
         cairo_runner.run_ended = true;
@@ -4834,7 +4827,7 @@ mod tests {
         Arc::get_mut(&mut program.shared_program_data).unwrap().data =
             vec_data![(1), (2), (3), (4), (5), (6), (7), (8)];
         //Program data len = 8
-        let mut cairo_runner = cairo_runner!(program, "all_cairo", true);
+        let mut cairo_runner = cairo_runner!(program, LayoutName::all_cairo, true);
         cairo_runner.program_base = Some(Relocatable::from((0, 0)));
         cairo_runner.execution_base = Some(Relocatable::from((1, 0)));
         cairo_runner.run_ended = true;
@@ -5346,14 +5339,15 @@ mod tests {
 
     #[test]
     fn get_cairo_pie_no_program_base() {
-        let runner = CairoRunner::new(&Program::default(), "all_cairo", false).unwrap();
+        let runner = CairoRunner::new(&Program::default(), LayoutName::all_cairo, false).unwrap();
         let vm = vm!();
         assert_eq!(runner.get_cairo_pie(&vm), Err(RunnerError::NoProgBase))
     }
 
     #[test]
     fn get_cairo_pie_no_execution_base() {
-        let mut runner = CairoRunner::new(&Program::default(), "all_cairo", false).unwrap();
+        let mut runner =
+            CairoRunner::new(&Program::default(), LayoutName::all_cairo, false).unwrap();
         let vm = vm!();
         runner.program_base = Some(Relocatable::from((0, 0)));
         assert_eq!(runner.get_cairo_pie(&vm), Err(RunnerError::NoExecBase))
@@ -5361,7 +5355,8 @@ mod tests {
 
     #[test]
     fn get_cairo_pie_no_segment_sizes() {
-        let mut runner = CairoRunner::new(&Program::default(), "all_cairo", false).unwrap();
+        let mut runner =
+            CairoRunner::new(&Program::default(), LayoutName::all_cairo, false).unwrap();
         let mut vm = vm!();
         runner.program_base = Some(Relocatable::from((0, 0)));
         runner.execution_base = Some(Relocatable::from((1, 0)));
@@ -5381,7 +5376,8 @@ mod tests {
 
     #[test]
     fn get_cairo_pie_ret_pc_segment_size_not_zero() {
-        let mut runner = CairoRunner::new(&Program::default(), "all_cairo", false).unwrap();
+        let mut runner =
+            CairoRunner::new(&Program::default(), LayoutName::all_cairo, false).unwrap();
         let mut vm = vm!();
         runner.program_base = Some(Relocatable::from((0, 0)));
         runner.execution_base = Some(Relocatable::from((1, 0)));
@@ -5403,7 +5399,8 @@ mod tests {
 
     #[test]
     fn get_cairo_pie_program_base_offset_not_zero() {
-        let mut runner = CairoRunner::new(&Program::default(), "all_cairo", false).unwrap();
+        let mut runner =
+            CairoRunner::new(&Program::default(), LayoutName::all_cairo, false).unwrap();
         let mut vm = vm!();
         runner.program_base = Some(Relocatable::from((0, 1)));
         runner.execution_base = Some(Relocatable::from((1, 0)));
@@ -5425,7 +5422,8 @@ mod tests {
 
     #[test]
     fn get_cairo_pie_execution_base_offset_not_zero() {
-        let mut runner = CairoRunner::new(&Program::default(), "all_cairo", false).unwrap();
+        let mut runner =
+            CairoRunner::new(&Program::default(), LayoutName::all_cairo, false).unwrap();
         let mut vm = vm!();
         runner.program_base = Some(Relocatable::from((0, 0)));
         runner.execution_base = Some(Relocatable::from((1, 1)));
@@ -5447,7 +5445,8 @@ mod tests {
 
     #[test]
     fn get_cairo_pie_ret_fp_offset_not_zero() {
-        let mut runner = CairoRunner::new(&Program::default(), "all_cairo", false).unwrap();
+        let mut runner =
+            CairoRunner::new(&Program::default(), LayoutName::all_cairo, false).unwrap();
         let mut vm = vm!();
         runner.program_base = Some(Relocatable::from((0, 0)));
         runner.execution_base = Some(Relocatable::from((1, 0)));
@@ -5469,7 +5468,8 @@ mod tests {
 
     #[test]
     fn get_cairo_pie_ret_pc_offset_not_zero() {
-        let mut runner = CairoRunner::new(&Program::default(), "all_cairo", false).unwrap();
+        let mut runner =
+            CairoRunner::new(&Program::default(), LayoutName::all_cairo, false).unwrap();
         let mut vm = vm!();
         runner.program_base = Some(Relocatable::from((0, 0)));
         runner.execution_base = Some(Relocatable::from((1, 0)));
@@ -5491,7 +5491,8 @@ mod tests {
 
     #[test]
     fn get_cairo_pie_ok() {
-        let mut runner = CairoRunner::new(&Program::default(), "all_cairo", false).unwrap();
+        let mut runner =
+            CairoRunner::new(&Program::default(), LayoutName::all_cairo, false).unwrap();
         let mut vm = vm!();
         runner.program_base = Some(Relocatable::from((0, 0)));
         runner.execution_base = Some(Relocatable::from((1, 0)));
@@ -5515,7 +5516,7 @@ mod tests {
             program_content,
             &CairoRunConfig {
                 proof_mode: true,
-                layout: "all_cairo",
+                layout: LayoutName::all_cairo,
                 ..Default::default()
             },
             &mut BuiltinHintProcessor::new_empty(),