+++ vm/src/vm/vm_core.rs
@@ -1245,6 +1245,7 @@ mod tests {
     use super::*;
     use crate::felt_hex;
     use crate::stdlib::collections::HashMap;
+    use crate::types::layout_name::LayoutName;
     use crate::types::program::Program;
     use crate::vm::runners::builtin_runner::{
         BITWISE_BUILTIN_NAME, EC_OP_BUILTIN_NAME, HASH_BUILTIN_NAME,
@@ -4310,7 +4311,7 @@ mod tests {
         .unwrap();
 
         let mut hint_processor = BuiltinHintProcessor::new_empty();
-        let mut cairo_runner = cairo_runner!(program, "all_cairo", false);
+        let mut cairo_runner = cairo_runner!(program, LayoutName::all_cairo, false);
         let mut vm = vm!();
 
         let end = cairo_runner.initialize(&mut vm, false).unwrap();
@@ -4335,7 +4336,7 @@ mod tests {
         .unwrap();
 
         let mut hint_processor = BuiltinHintProcessor::new_empty();
-        let mut cairo_runner = cairo_runner!(program, "all_cairo", false);
+        let mut cairo_runner = cairo_runner!(program, LayoutName::all_cairo, false);
         let mut vm = vm!();
 
         let end = cairo_runner.initialize(&mut vm, false).unwrap();